User mie (#18)
* let's start with a new cloud repository

* recover emacs files added to gitignore

* bugfix with NWPSAF inputs and related cleanup

* read mie data and compute betal

* Extra checks in rttov, still commented

* improve checks on netcdf inputs

* Optical properties implemented

* reshaped rttov

* reshaped rttov

* documented

* doxygen documentation implemented

* implement github action

* update doxygen

* update doxygen

* update doxygen

* update doxygen

* update doxygen

* remove verbose

+++ src/utils/utils_phys.f90
@@ -254,7 +254,7 @@ contains
 
     logical   PASS1
     integer(kind=4)   DELTA, LEAP
-    double precision  DEC, DEN, ECLONG, GMST, HA, JD, LMST, &
+    real(wp)  DEC, DEN, ECLONG, GMST, HA, JD, LMST, &
          MNANOM, MNLONG, NUM, OBLQEC, PI, RA, &
          RPD, REFRAC, TIME, TWOPI
